End of Phase 2

+++ Backend/src/main/java/com/swapitServer/product/ProductRepository.java
@@ -6,7 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface ProductRepository extends JpaRepository<Product, Long> {
 
     Product findByName(String name);
-    Product findByPrice(long price);
-    //Product findByIdd(long id);
+    Product findByPrice(float price);
     Product findById(long id);
 }
